refactor: avoid duplicate code between user roles and user groups tabs
- Extract duplicate functions
- Refactor to use generic parameter
